app/requirements.txt
numpy==1.26.4
fastapi
uvicorn[standard]
boto3
python-multipart
pillow
opencv-python-headless

# PyTorch (CPU 전용)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
